Move ReservationUnitQuerySet to its own file
